added am income source table

+++ src/Services/MerchantService.cs
@@ -106,7 +106,7 @@ public class MerchantService : IMerchantService
         return expenses.GroupBy(e => e.MerchantId)
                     .Select(g =>
                     {
-                        var results = g.Aggregate(new MerchantListItemAccumulator(g.Key, merchants, categories), (acc, e) => acc.Accumulate(e), acc => acc.Compute());
+                        var results = g.Aggregate(new MerchantListItemAggregator(g.Key, merchants, categories), (acc, e) => acc.Accumulate(e), acc => acc.Compute());
                         return new MerchantListItem()
                         {
                             Id = g.Key.HasValue ? g.Key.Value : 0,
@@ -138,17 +138,17 @@ public class MerchantService : IMerchantService
             Notes = merchant.Notes,
             IsOnline = merchant.IsOnline,
 
-            ExpenseTotals = expenses.Aggregate(new ExpenseTotalsAggregator(),
+            ExpenseTotals = expenses.Aggregate(new TotalsAggregator<ExpenseEntity>(),
                 (acc, e) => acc.Accumulate(e),
                 acc => acc.Compute()),
 
             MostUsedCategory = GetExpenseCategoryOccurances(subCategories, expenses).FirstOrDefault().Key,
 
             AnnualExpenseStatistics = expensesSpanMultipleYears ?
-            GetAnnualExpenseStatistics(expenses) : new List<ExpenseStatistics>(),
+            AggregateUtilities<ExpenseEntity>.GetAnnualStatistics(expenses) : new List<AnnualStatistics>(),
 
             MonthlyExpenseStatistics = expensesSpanMultipleYears ?
-            new List<MonthlyExpenseStatistics>() : GetMonthlyExpenseStatistics(expenses),
+            new List<MonthlyStatistics>() : AggregateUtilities<ExpenseEntity>.GetMonthlyStatistics(expenses),
 
             PurchaseCategoryOccurances = GetExpenseCategoryOccurances(subCategories, expenses),
             PurchaseCategoryTotals = GetExpenseCategoryTotals(subCategories, expenses),
@@ -190,68 +190,6 @@ public class MerchantService : IMerchantService
                 Sum = x.Expenses.Sum(e => e.Amount)
             }).Where(x => x.Sum > 0).OrderByDescending(x => x.Sum).ToDictionary(k => k.Category, v => v.Sum);
     }
-    private List<ExpenseStatistics> GetAnnualExpenseStatistics(ExpenseEntity[] expenses)
-    {
-        return expenses.GroupBy(e => e.Date.Year)
-                        .Select(g =>
-                        {
-                            var results = g.Aggregate(new ExpenseStatisticsAggregator(),
-                                (acc, e) => acc.Accumulate(e),
-                                acc => acc.Compute());
-                            return new ExpenseStatistics()
-                            {
-                                Year = g.Key,
-                                Average = results.Average,
-                                Min = results.Min,
-                                Max = results.Max,
-                                Total = results.Total,
-                                Count = results.Count
-                            };
-                        }).OrderBy(x => x.Year).ToList();
-    }
-    private List<MonthlyExpenseStatistics> GetMonthlyExpenseStatistics(ExpenseEntity[] expenses)
-    {
-        var monthlyExpenseStatistics = new List<MonthlyExpenseStatistics>();
-
-        var monthlyStats = expenses.GroupBy(e => e.Date.DateTime)
-        .Select(g =>
-        {
-            var results = g.Aggregate(new ExpenseStatisticsAggregator(),
-                (acc, e) => acc.Accumulate(e),
-                acc => acc.Compute());
-
-            return new MonthlyExpenseStatistics()
-            {
-                Date = g.Key,
-                Average = results.Average,
-                Min = results.Min,
-                Max = results.Max,
-                Total = results.Total,
-                Count = results.Count
-            };
-        }).OrderBy(x => x.Date).ToList();
-        var year = expenses.FirstOrDefault().Date.Year;
-        for (var i = 1; i <= 12; i++)
-        {
-            if (monthlyStats.Any(x => x.Date.Month == i))
-            {
-                monthlyExpenseStatistics.Add(monthlyStats.Where(x => x.Date.Month == i).FirstOrDefault());
-            }
-            else
-            {
-                monthlyExpenseStatistics.Add(new MonthlyExpenseStatistics()
-                {
-                    Date = new DateTime(year, i, 1),
-                    Average = 0,
-                    Min = 0,
-                    Max = 0,
-                    Total = 0,
-                    Count = 0
-                });
-            }
-        }
-        return monthlyExpenseStatistics;
-    }
 
     public async Task<int> CreateMerchantAsync(Merchant request)
     {
@@ -267,15 +205,25 @@ public class MerchantService : IMerchantService
     {
         var merchants = await _merchantRepo.Find(x => x.Name == request.Name);
         if (merchants.Any(x => x.Id != request.Id))
-            throw new DuplicateNameException(nameof(MerchantEntity), request.Name);
+            throw new DuplicateNameException(request.Name, nameof(MerchantEntity));
 
-        var merchant = _mapper.Map<MerchantEntity>(request);
+        var merchant = merchants.FirstOrDefault(x => x.Id == request.Id.Value);
+
+        merchant.Name = request.Name;
+        merchant.City = request.IsOnline ? null : request.City;
+        merchant.State = request.IsOnline ? null : request.State;
+        merchant.Notes = request.Notes;
+        merchant.IsOnline = request.IsOnline;
+        merchant.SuggestOnLookup = request.SuggestOnLookup;
+
         return await _merchantRepo.Update(merchant);
     }
 
     public async Task<bool> DeleteMerchantAsync(int id)
     {
         var merchant = await _merchantRepo.FindById(id);
+        if (merchant == null)
+            throw new MerchantNotFoundException(id.ToString());
         var expenses = await _expenseRepo.Find(x => x.MerchantId == id);
         if (!expenses.Any())
             return await _merchantRepo.Delete(merchant);
